+++ server/apps/api/src/app/auth/auth.controller.ts
@@ -6,7 +6,6 @@ import { ApiErrorResponse } from '../../common/decorator/api-error-response.deco
 import { ApiSuccessResponse } from '../../common/decorator/api-success-response.decorator';
 import { InvalidPhoneCodeException } from '../../common/exception/InvalidPhoneCodeException';
 import { UserNotFoundException } from '../../common/exception/UserNotFoundException';
-import { TokenService } from '../../common/token/token.service';
 import { AuthService } from './auth.service';
 import { RegisterRequest } from './dto/register-request.dto';
 import { SendPhoneCodeDto } from './dto/send-phone-code.dto';
@@ -15,7 +14,7 @@ import { TokenResponse } from './dto/token-response.dto';
 @Controller('/auth')
 @ApiTags('Auth')
 export class AuthController {
-  constructor(private readonly authService: AuthService, private readonly tokenService: TokenService) {}
+  constructor(private readonly authService: AuthService) {}
 
   @Post('/phone/code')
   @ApiSuccessResponse(HttpStatus.NO_CONTENT)
@@ -27,8 +26,7 @@ export class AuthController {
   @ApiSuccessResponse(HttpStatus.CREATED, TokenResponse)
   @ApiErrorResponse(InvalidPhoneCodeException)
   async register(@Body() body: RegisterRequest) {
-    const user = await this.authService.register(body.phone, body.code);
-    const tokens = await this.tokenService.generateAuthToken(user);
+    const tokens = await this.authService.register(body.phone, body.code);
 
     return ResponseEntity.OK_WITH_DATA(new TokenResponse(tokens));
   }
@@ -37,8 +35,7 @@ export class AuthController {
   @ApiSuccessResponse(HttpStatus.OK, TokenResponse)
   @ApiErrorResponse(InvalidPhoneCodeException, UserNotFoundException, BadParameterException)
   async login(@Body() body: RegisterRequest) {
-    const user = await this.authService.login(body.phone, body.code);
-    const tokens = await this.tokenService.generateAuthToken(user);
+    const tokens = await this.authService.login(body.phone, body.code);
 
     return ResponseEntity.OK_WITH_DATA(new TokenResponse(tokens));
   }
